Logoコンポーネント追加、Headerコンポーネント調整
