--- src/components/Atoms/Header/index.tsx
@@ -1,15 +1,0 @@
-import Styles from './styles.module.css';
-
-const Title: React.FC = () => {
-  return (
-    <header className={Styles.header}>
-      <a href="/" className={Styles.logo}>
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 280 56">
-          <path id="logo" d="M15.536-1.447c7.7,0,12.357-4.542,12.357-11.221,0-6.626-5.3-9.191-10.211-11.435l-2.73-1.229c-4.368-1.977-8.518-4.008-8.518-8.6,0-3.74,2.894-6.84,7.972-6.84,4.532,0,8.463,2.511,10.866,4.969l.874-2.992A24.523,24.523,0,0,0,15.007-41.63c-7.59,0-12.558,4.382-12.558,10.26,0,6.359,5.733,9.137,10.047,11.061l2.73,1.176c4.041,1.817,8.736,4.115,8.736,9.244,0,4.382-3.6,7.588-8.9,7.588-4.423,0-9.61-2.191-12.231-5.344L1.848-4.279A38.912,38.912,0,0,0,15.536-1.447ZM35.428-43.767V-2.088h5.133V-45.37h-.218ZM46.4-20.523l.983-.855A98.671,98.671,0,0,1,59.4-30.355v-.107H55.631a80.2,80.2,0,0,1-8.845,8.282l-4.8,4.061L59.289-.912h.273V-7.7ZM85.88-30.462c-.328,3.26-1.8,8.5-4.75,15.5L78.29-8.34,68.462-30.462H62.893L75.615-1.928l-1.8,4.275a8.375,8.375,0,0,1-7.7,5.4,9.243,9.243,0,0,1-2.675-.374l-.328,3.046c4.914,0,9.118-1.87,11.739-8.015l7.207-16.992c2.948-6.947,4.7-11.275,6.989-15.87ZM125.958-1.447a18.877,18.877,0,0,0,16.6-9.3l-.655-.374A15.815,15.815,0,0,1,128.415-3.8c-8.791,0-15.4-6.947-15.4-18.756,0-10.42,5.46-18.221,13.814-18.221,5.3,0,9.665,3.26,12.449,7.8l2.73-2.4A20.12,20.12,0,0,0,127.1-41.63c-11.521,0-20.2,7.962-20.2,20.412C106.9-9.195,115.037-1.447,125.958-1.447Zm35.983,0c8.518,0,14.906-5.931,14.906-14.962,0-9.084-6.115-14.695-14.415-14.695-8.518,0-14.906,5.985-14.906,15.015C147.526-7,153.641-1.447,161.94-1.447Zm.328-.748c-5.406,0-9.173-5.664-9.173-14.214,0-8.443,3.549-13.947,9.009-13.947,5.406,0,9.173,5.718,9.173,14.267C171.277-7.7,167.728-2.195,162.268-2.195Zm34.891-41.573A13.834,13.834,0,0,1,203-42.485l.382-.641a16.776,16.776,0,0,0-8.572-2.458c-5.679,0-9.391,3.473-9.391,10.1v5.023h-4.2v1.176h4.2v27.2h5.187v-27.2h8.409l.764-1.176h-9.173v-5.29C190.606-41.256,193.063-43.767,197.159-43.767Zm21.9,0a13.834,13.834,0,0,1,5.842,1.282l.382-.641a16.776,16.776,0,0,0-8.572-2.458c-5.679,0-9.391,3.473-9.391,10.1v5.023h-4.2v1.176h4.2v27.2H212.5v-27.2h8.409l.764-1.176H212.5v-5.29C212.5-41.256,214.959-43.767,219.054-43.767Zm30.522,25.489c.109-7.534-3.931-12.824-11.794-12.824-8.409,0-14.688,6.038-14.688,15.282,0,8.656,5.788,14.374,13.6,14.374,5.133,0,9.446-2.4,12.558-7.053l-.655-.374a12.094,12.094,0,0,1-9.664,5.13c-5.9,0-10.32-4.7-10.32-13.573v-.962ZM237.345-30.3c4.641,0,7.59,4.435,7.262,11.168H228.664C229.21-25.653,232.6-30.3,237.345-30.3Zm44.5,12.023c.109-7.534-3.931-12.824-11.794-12.824-8.409,0-14.688,6.038-14.688,15.282,0,8.656,5.788,14.374,13.6,14.374,5.133,0,9.446-2.4,12.558-7.053l-.655-.374a12.094,12.094,0,0,1-9.664,5.13c-5.9,0-10.32-4.7-10.32-13.573v-.962ZM269.615-30.3c4.641,0,7.59,4.435,7.262,11.168H260.933C261.479-25.653,264.865-30.3,269.615-30.3Z" transform="translate(-1.848 45.584)"/>
-        </svg>
-      </a>
-    </header>
-  )
-}
-
-export default Title;
