+++ src/components/Atoms/Logo/index.stories.tsx
@@ -1,0 +1,8 @@
+import Header from './index';
+
+export default {
+  title: 'Atoms/Logo',
+  component: Header,
+};
+
+export const $default: React.FC = () => <Header />;
