+++ src/components/Atoms/Logo/styles.module.css
@@ -1,0 +1,4 @@
+.logo {
+  width: 100%;
+  fill: currentColor;
+}
